splat: generalise use of tamfits BasicHDU.getBlankValue
The method nom.tam.fits.BasicHDU.getBlankValue returns an int
in the starjava version of nom.tam.fits and a long in more
recent versions.  The code can easily be changed so it works
with both.

+++ splat/src/main/uk/ac/starlink/splat/data/FITSSpecDataImpl.java
@@ -648,7 +648,7 @@ public class FITSSpecDataImpl
             double bzero = hdurefs[hdunum].getBZero();
 
             //  Get the BLANK value, note if none present.
-            int blank = 0;
+            long blank = 0;
             boolean haveblank = true;
             try {
                 blank = hdurefs[hdunum].getBlankValue();
